RT-32589: Add better diagnostic and debug tools for dirty regions, culling, and overdraw Reviewed-By: Steve Northover, Martin Sladecek
Added a new dirty bit for debug information that needs to be sent down. Also completely changed the way we draw in ViewPainter, specifically when dirty regions are used and even more specifically when show dirty and show overdraw are turned on. Added some new prism settings for showing overdraw, and printing the render graph to the pulse logger. Also replaced the drawCullBits to instead draw overdraw boxes. You can get a better sense of cull bits by looking at the output of the render graph being printed to the pulse logger.